#7 removing mock data for leaderboard into a separate file
