+++ packages/client/src/pages/LeaderboardPage/LeaderboardPage.tsx
@@ -1,14 +1,9 @@
+import { Table, Typography } from 'antd';
 import { Layout } from 'layouts/Layout';
 import { en } from 'translations';
-import { Table, Typography } from 'antd';
+import { data } from './mock';
 
 import './Leaderboard.scss';
-
-interface IData {
-  name: string;
-  score: string;
-  key: number;
-}
 
 const columns = [
   {
@@ -16,31 +11,8 @@ const columns = [
     dataIndex: 'name',
   },
   {
-    title: 'Очки',
-    dataIndex: en['leaderboard.table.score'],
-  },
-];
-
-const data: IData[] = [
-  {
-    name: 'Example 1',
-    score: '5',
-    key: 1,
-  },
-  {
-    name: 'Example 2',
-    score: '4',
-    key: 2,
-  },
-  {
-    name: 'Example 3',
-    score: '3',
-    key: 3,
-  },
-  {
-    name: 'Example 4',
-    score: '2',
-    key: 4,
+    title: en['leaderboard.table.score'],
+    dataIndex: 'score',
   },
 ];
