+++ packages/client/src/pages/LeaderboardPage/mock.ts
@@ -1,0 +1,28 @@
+interface IData {
+  name: string;
+  score: string;
+  key: number;
+}
+
+export const data: IData[] = [
+  {
+    name: 'Example 1',
+    score: '5',
+    key: 1,
+  },
+  {
+    name: 'Example 2',
+    score: '4',
+    key: 2,
+  },
+  {
+    name: 'Example 3',
+    score: '3',
+    key: 3,
+  },
+  {
+    name: 'Example 4',
+    score: '2',
+    key: 4,
+  },
+];
